Move trinity.db.* to trinity.db.eth.*
