+++ trinity/plugins/builtin/json_rpc/plugin.py
@@ -14,7 +14,7 @@ from trinity.config import (
     TrinityConfig
 )
 from trinity.chains.base import BaseAsyncChain
-from trinity.db.manager import (
+from trinity.db.eth1.manager import (
     create_db_manager
 )
 from trinity.extensibility import (
